added date comparison to schedule and updated readme

+++ src/main/java/com/example/urldownloadtool/scheduled/ScheduledTasks.java
@@ -6,6 +6,7 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 import java.io.File;
+import java.util.Date;
 
 @Component
 public class ScheduledTasks {
@@ -14,8 +15,13 @@ public class ScheduledTasks {
     private static final String userDir = System.getProperty("user.dir");
     private static final String videoFolder= userDir+ "/videos/";
 
+    Date currentDate = new Date();
+    long currentDateInMilliSeconds = currentDate.getTime();
+
+
+
     // This scheduled Task runs every 10 Minutes and deletes videos from the folder
-    @Scheduled(fixedRate = 600000)
+    @Scheduled(fixedRate = 60000)
     public void cleanUpVideos(){
         File videoDir = new File(videoFolder);
 
@@ -27,8 +33,12 @@ public class ScheduledTasks {
             for (File file : videosInFolder
             ) {
                 if (file.isFile()) {
-                    log.info("current File deleted:" + file.toString());
-                    file.delete();
+                    // compare current date and last modified date of file and delete it when it's older than 5 minutes
+                    long fileDateInMilliseconds = file.lastModified();
+                    if (currentDateInMilliSeconds - fileDateInMilliseconds > 300000) {
+                        log.info("currentDate: "+ currentDateInMilliSeconds + " And file date: "+ fileDateInMilliseconds);
+                        file.delete();
+                    }
                 }
 
             }
